Inclusão de permissões para nível CS

+++ sermil-core/src/main/java/br/mil/eb/sermil/core/servicos/CidFotoServico.java
@@ -11,35 +11,47 @@ import br.mil.eb.sermil.core.dao.CidFotoDao;
 import br.mil.eb.sermil.core.exceptions.SermilException;
 import br.mil.eb.sermil.modelo.CidFoto;
 
-/** Serviço de Foto de Cidadão.
+/** Serviço de arquivo de foto de um cidadão.
  * @author Abreu Lopes
  * @since 4.0
- * @version $Id: CidFotoServico.java 2427 2014-05-15 13:23:38Z wlopes $
+ * @version 5.3.1
  */
 @Named("cidFotoServico")
 public class CidFotoServico {
 
-  protected static final Logger logger = LoggerFactory.getLogger(CidFotoServico.class);
+   protected static final Logger logger = LoggerFactory.getLogger(CidFotoServico.class);
 
-  @Inject
-  private CidFotoDao cidFotoDao;
+   @Inject
+   private CidFotoDao cidFotoDao;
 
-  public CidFotoServico() {
-    logger.debug("CidFotoServico iniciado");
-  }
+   public CidFotoServico() {
+      logger.debug("CidFotoServico iniciado");
+   }
 
-  public CidFoto obter(final Long ra) throws SermilException {
-    return this.cidFotoDao.findById(ra);
-  }
+   public CidFoto obter(final Long ra) throws SermilException {
+      if (ra == null) {
+         throw new SermilException("Informe um RA válido.");
+      }
+      logger.debug("RA = {}", ra);
+      return this.cidFotoDao.findById(ra);
+   }
 
-  @Transactional
-  public void excluir(final CidFoto foto) throws SermilException {
-    this.cidFotoDao.delete(foto);
-  }
+   @Transactional
+   public void excluir(final CidFoto foto) throws SermilException {
+      if (foto == null) {
+         throw new SermilException("Foto não existe.");
+      }
+      logger.debug("Foto = {}", foto);
+      this.cidFotoDao.delete(foto);
+   }
 
-  @Transactional
-  public void salvar(final CidFoto foto) throws SermilException {
-    this.cidFotoDao.save(foto);
-  }
+   @Transactional
+   public void salvar(final CidFoto foto) throws SermilException {
+      if (foto == null) {
+         throw new SermilException("Foto não existe.");
+      }
+      logger.debug("Foto = {}", foto);
+      this.cidFotoDao.save(foto);
+   }
 
 }
